+++ packages/datetime/tsup.config.ts
@@ -1,0 +1,4 @@
+import { defineConfig } from 'tsup';
+import { tsupCommonConfig } from '../../utils/tsup.common';
+
+export default defineConfig(tsupCommonConfig(import.meta.dir));